// frontend/src/api/types.ts
export type Visibility = 'PUBLIC' | 'PRIVATE'

export type TransferResponse = { transactionId: string; payer: number; payee: number; value: number; createdAt: string }